MVC Model - UserBusiness - View - Create Order
Current Progress:
users are able to login and i have made progress in the development of the business user functionality. I have created a new controller for the business user and a new view for the business user. I have also created a new service.
The user is able to create an order successfully which is populated on the database.

The next steps are to create the functionality for the business user to view the orders that have been placed and to update orders.

I have made efforts to restructure the code base to make it more modular and easier to maintain. I have created a new folder structure that will separate the different components of the project.